requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.37.23
botocore==1.37.23
cachetools==5.5.2
cassandra-driver==3.29.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
geomet==0.2.1.post1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
numpy==2.2.4
pandas==2.2.3
proto-plus==1.26.1
protobuf==5.29.4
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pyspark==3.5.5
pyspark-stubs==3.0.0.post3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
